@@ -40,7 +40,7 @@ else
 <div id="PizzaSummaryId">
     @foreach (var pizza in Model.PizzaList)
     {
-        @Html.Partial("PizzaSummary", pizza)
+        @await Html.PartialAsync("PizzaSummary", pizza)
     }
 </div>
 <br />
