@@ -10,7 +10,7 @@
 
 @foreach (var pizza in Model)
 {
-    @Html.Partial("PizzaSummary", pizza)
+    @await Html.PartialAsync("PizzaSummary", pizza)
 }
 <br />
     
