@@ -10,6 +10,6 @@
 
 @foreach (var pizza in Model)
 {
-    @Html.Partial("PizzaSummary", pizza)
+    @await Html.PartialAsync("PizzaSummary", pizza)
 }
 <br />
