@@ -1,3 +1,3 @@
 @model IEnumerable<Reviews>
 
-@Html.Partial("ListAllReviewsPartial", Model)
+@await Html.PartialAsync("ListAllReviewsPartial", Model)
