@@ -30,7 +30,7 @@
 </div>
 
 <div id="defaultResult">
-    @Html.Partial("ListAllReviewsPartial", Model)
+    @await Html.PartialAsync("ListAllReviewsPartial", Model)
 </div>
 
 @section Scripts
